+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 0.10.0
+### Added
+- DeepSeq instances for all data types.
+
+### Changed
+- There is no `SockAddr` inside `NetworkAddress` anymore.
+
 ## 0.9.8
 ### Added
 - Ord instance for `DerivPathI`